Fix compatibility with oslo.db 12.1.0
oslo.db 12.1.0 has changed the default value for the 'autocommit'
parameter of 'LegacyEngineFacade' from 'True' to 'False'. This is a
necessary step to ensure compatibility with SQLAlchemy 2.0. However, we
are currently relying on the autocommit behavior and need changes to
explicitly manage sessions. Until that happens, we need to override the
default.

This also moves the watcher queue declaration from the pipeline level
(where it is no longer valid) to the project level.

Co-Authored-By: Thierry Carrez <thierry@openstack.org>
Co-Authored-By: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 5d70c207cddbaedceba6e1e6e9ae89bbd851aacc)
Change-Id: I7db39d958d087322bfa0aad70dfbd04de9228dd7