Drop nova command reference from the code
In DevStack environment, nova service-list command does not
exist. Distro suggests to install python-novaclient from package.

In Strategies documentation, we generate the docs from following
code.[1]
```
       * - ``migration``
         - .. watcher-term:: watcher.applier.actions.migration.Migrate
       * - ``change_nova_service_state``
         - .. watcher-term:: watcher.applier.actions.change_nova_service_state.ChangeNovaServiceState
```
and with in code, we use nova python binding to get list services[2]
and we are not calling openstack cli reference with in the code.

Documenting the equivalent openstack command does not seems to be useful
in the help text as we are using python binding.

Links:
[1]. https://github.com/openstack/watcher/blob/c4acce91d6bb87b4ab865bc8e4d442a148dba1d5/doc/source/strategies/host_maintenance.rst?plain=1#L45
[2]. https://github.com/openstack/watcher/blob/c4acce91d6bb87b4ab865bc8e4d442a148dba1d5/watcher/common/nova_helper.py#L150-L152

Change-Id: I0c663c9741fae94bdb9c30f46d3d396325a33948
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>