Add unit test to check action plan state when a nested action fails
This patch is adding a new unit test to check the behavior of the action
plan when one of the actions in it fails during execution.

Note this is to show a bug, and the expected state will be changed in
the fixing patch.

Related-Bug: #2106407
Change-Id: I2f3fe8f4da772a96db098066d253e5dee330101a
(cherry picked from commit b36ba8399e4be1362ce41f8809c4f43df05a4dc6)
Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>