Ignore cinder notifications if the Storage model is empty
The cinder collector has some logic to skip creating the model when
there are not audits which require the information. In this case,
when processing cinder notifications as `volume.create.end` it fails
because the model is empty.

In this case, the desired behavior is to just ignore the messages until
a first audit is executed and a model is created. Since then, we want to
keep the model updated when notifications are enabled.

This patch making the cinder notifications endpoints to ignore messages
in that case. This is similar to the behavior of the nova endpoint [1].

[1] https://opendev.org/openstack/watcher/src/commit/355671e9791d20acba21ef4af64d42c658e00276/watcher/decision_engine/model/notification/nova.py#L361-L363

Closes-Bug: #2120620
Change-Id: I5fee783efb0e588235c75a395f39e4baedbc4e7a
Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>