Add some debug logs to storage model
Some functions related to the storage model log exceptions in a way that
is not clear for the user where the problem could come from. Sometimes,
the exception comes from a check, but it's not an actual error. This
patch adds some debug logs and improve some exception message to make
clearer in the logs when an error is important and when is just the
result of a check.

Change-Id: Ia2c7d50c44a1740ec3f4f5c7454daceb95f6e1b2
Signed-off-by: jgilaber <jgilaber@redhat.com>