APISchedulingService migrate audits also on first discovery of services
Currently, the APISchedulingService only migrates ongoing audits
when it detects the decision-engine changes the state from ACTIVE
to FAILED. That makes that, audits assigned to a FAILED decision-engine
are not reassigned when the watcher-api service is started after the
decision-engine dies.

After this patch, the APISchedulingService will also migrate ongoint audits
also when a decision-engine is detected as FAILED when the service is
started.

I'm also moving the audit migration logic to a separate function to
improve readability.

Closes-Bug: #2127777

Change-Id: Ic3c9d89973ca976dde9f6e9b423671e52b049cdb
Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>