use the correct package name for microversion-parse
As of Ib97eea1037fac541c8ebadc526ae9c3bb160f685 the
requirements check job nolonger uses pkg_resources
to normalise the package names, this highlighted that
we are not using the actuall name and were relying on
that normalization for microversion-parse.

This change fixes that by using the correct name.

Change-Id: I5fb1717b6ce3662edf35a1ced10bc5fcbcd8fa07
Signed-off-by: Sean Mooney <work@seanmooney.info>